fix: don't accumulate result state with rhel provider (#253)
* fix: don't accumulate result state with rhel provider

Sets the result store policy for the `rhel` provider to
`DELETE_BEFORE_WRITE`.  There shouldn't be a reason for the rhel
provider to need to accumulate result state between runs as it always
iterates over the full input set.  This will address an issue where
vulnerabilities marked as `not-affected` don't get dropped from the
results.

Signed-off-by: Weston Steimel <weston.steimel@anchore.com>

* fix: merge out of support entries when affected or fixed version in more recent release

Signed-off-by: Weston Steimel <weston.steimel@anchore.com>

* chore: bump label dataset

Signed-off-by: Weston Steimel <weston.steimel@anchore.com>

---------

Signed-off-by: Weston Steimel <weston.steimel@anchore.com>