fix(wolfi): update vulnerability reference links
The logic that generates the reference link for wolfi advisories was
not updated to account for GitHub Security Advisories, resulting in
invalid links for those records.  Also, the urls generated for CVE ids
are also no longer valid as they still pointed to the old mitre website
which is no longer updated.

Signed-off-by: Weston Steimel <weston.steimel@proton.me>