hack: hard-code severity for debian CVE-2023-44487
The Debian feed contains per-package urgency for CVEs, which Vunnel
translates into per-package severity, and the collapses into a single
per-CVE serverity by having the most severe severity win. However,
unknown severities always lose the comparison. This means that the
particular CVE here, CVE-2023-44487, which has a single negligible
package and a bunch of unknowns, is treated as negligible severity for
every package, which is incorrect. Hard-code it to high severity while
we figure out what schema changes and logic changes are needed put in a
lasting fix.

Signed-off-by: Will Murphy <will.murphy@anchore.com>