fix: special case "greater than version zero" test
Vulnerability criteria in the SLES parser that have an RPM version of
"0:0-0" and an operator of "greater than" mean that any version of the
package is vulnerable (because all packages have a version greater than
zero). Therefore, emit a FixedIn with version of "None" in this case, to
tell grype DB that there is no fix for this package yet.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>