fix: stop module versions sneaking in
Previously, both the module and package PURLs might match normalized
package name, resulting in an incorrect version being returned. Instead,
keep the remediation product ids in a list, not a set, so that they can
be iterated in a deterministic sequence, and never return a module purl
as though it were a package.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>