feat: add skip download CLI option (#802)
This adds the ability to skip downloading to the CLI, so that users can
run "vunnel run <provider> --skip-download". This is useful when working
on a provider with a slow download step.

Currently, only Rocky Linux provider supports this. Other providers
passed this flag will log a warning and then ignore the flag.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>