Revert "chore: remove redundant release check (#820)" (#821)
* Revert "chore: remove redundant release check (#820)"

This reverts commit 1f2fac203aa199e61a5c0908a56c9f3c000deada. The first
release after that commit was made failed with a permissions error in
the step that was changed. Therefore, reverting the commit on the
assumption that it was to blame. This will have the side effect of
causing there to once again be two manual confirmation steps in a vunnel
release, but this is preferable to partially executed releases.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>

* document why environment key is needed

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>

---------

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>