feat: optionally download alma data during rhel run
AlmaLinux appears not to have its own disclosures feed, but only
advisories, which seem to correspond 1:1 with Red Hat advisories.
Therefore, change the rhel provider to optionally download Alma
advisories and emit copies of rhel vulnerabilities with the fix info
changed to use the versions and URLs from Alma advisories.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>