chore: avoid needless compression (#841)
Previously, grype could only import a compressed vulnerability database
(the result of 'grype-db package'). However, now that grype can import
an uncompressed database (the result of 'grype-db build' without needing
to package), skip the costly compression step for databases that will
not be distributed, such as those that will be reused on the same host
in CI or built during local development.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>