use a consensus version where available
Sometimes, every package on an advisory is fixed at the same version,
called here a "consensus version". When emitting the Alma version of a
Red Hat vulnerability, if there is an AlmaLinux advisory that omits the
specific package, before falling back to the fixed version from the RHEL
advisory, check whether there's consensus in the version on the Alma
advisory. If so, assume that the package in question is also fixed at
that version.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>