feat(nvd): allow synthesis of new NVD records from overrides
Sometimes it is necessary to get information about a CVE out before the
base json record has been imported to NVD.  This allows synthesis of a
record from information contained in the override, but only when an NVD
record does not exist.  Once a record exists, the override will function
as normal and ONLY the cpe configurations will be considered.

Signed-off-by: Weston Steimel <author@code.w.steimel.me.uk>