fix: infer missing vulns from extended support fixes
Previously, it was possible for Vunnel to emit a vuln that meant,
essentially, "Red Hat patched this in RHEL 6 ELS" without recording that
it was therefore vulnerable in regular RHEL 6. For an example see
https://access.redhat.com/security/cve/cve-2021-25220
which has 'bind' listed as fixed in a RHEL 6 ELS release, but no record
that bind was affected in regular RHEL 6. Infer the existence of the
missing RHEL 6 vuln to prevent false negatives.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>