ci: filter by tag in make all-providers (#960)
* ci: filter by tag in make all-providers

Previously, "make all-providers" printed all providers even if a tag was
passed, leading to duplicate CI runs.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>

* ci: fail on missing label if providers needed

This is yet another bug fix to the pr-quality-gate yaml. It prevents the
job from passing incorrectly when the providers were skipped due to
missing label instead of due to not being needed.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>

---------

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>