fix: no els records
A recent fix (572c538) added the inference of missing RHEL records.
Essentially, if Red Hat claim that RHEL 6 ELS fixes package foo for
CVE-1234, infer that package foo was vulnerable to CVE-1234 in non-ELS
RHEL 6. As a side effect of this change, a fix version scoped to RHEL 6
ELS was emitted. However, versions of Grype before 0.97.0, when the
channel column was added to the OS table, fail to import the resulting
grype db due to a uniqueness constraint on the OS table, since for
example RHEL 8 and RHEL 8 ELS look identical if there is no channel
column to put the ELS value in.

Therefore, instead of emitting both the inferred wont-fix value and the
ELS-only fixed version, emit only the inferred wont-fix value, so that
false negatives are fixed but the uniqueness constraint in old versions
of Grype is preserved.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>